backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy==2.0.23
passlib[bcrypt]
python-jose[cryptography] 